Add logging
We're going to use the winston logging library for logging, which will
eventually mean we can use services outside of AWS if we wanted to.  For
now we will use the built in CloudWatch which keeps an eye on any logged
output and makes those logs available via AWS console.

Issue #1