Move code to a src directory
This is going to make things like linting and packaging a bit easier in
the long run since the active code will be set aside in its own
directory.

This also fixes a few errors / mistakes with the original addition of
logging.  Because of the rename it was hard to merge this into the
previous patch.