Update dependencies for deployment
We want to use linting and a bundler in order to get this project ready
for controlled lambda deployment.  This sets all of that up and adds the
appropriate lines to gitignore to prevent storage of things like caches
or distribution packages.

Side note: this also removes .DS_Store from gitignore since it's best
practice to NOT ignore environment specific files like that in the
collective repository.  OSX developers on this project should instead
set up a global .gitignore on their development machine.  Here is
a handy guide: https://gist.github.com/subfuzion/db7f57fff2fb6998a16c